backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
sqlalchemy==2.0.31
boto3==1.34.131
python-dotenv==1.0.1
